requirements.txt
boto3>=1.28.0
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0
opentelemetry-exporter-otlp>=1.19.0
python-dotenv>=1.0.0
openinference-instrumentation
openinference-semantic-conventions
